refactor(d51): split USB into acm/midi/hid files, remove comments
- grid_d51_usb.c now only contains descriptors, IRQ handlers, and init
- grid_d51_usb_acm.{c,h}: CDC/ACM platform API and TinyUSB callbacks
- grid_d51_usb_midi.{c,h}: MIDI platform API and TinyUSB callbacks
- grid_d51_usb_hid.{c,h}: HID platform API, TinyUSB callbacks, websocket stubs
- Remove explanatory comments from atmel_start.c and atmel_start.h
- Matches ESP32-S3 per-interface file structure
