+++ d51n20a/atmel_start.c
@@ -5,6 +5,5 @@
  **/
 void atmel_start_init(void) {
   system_init();
-  // usb_init() removed: TinyUSB is now initialized via grid_d51_usb_init()
   stdio_redirect_init();
 }
